OnboardingReviewDTO e Nota Integer

+++ src/main/java/com/adriankdev/RecommendationAPI/DTO/ReviewCreateDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class ReviewCreateDTO {
-    private Double nota;
+    private Integer nota;
     private String comentario;
 }
